feat: Implement API key validity check and credit management services
Adds credit management and API key validity services

Introduces interfaces and implementations for managing application credit limits and utilization using Redis.
Enhances Tenet-related services to support plan credit limit retrieval.
Refactors code structure for better separation of concerns and removes unused classes to streamline cache logic.

+++ TenetService/src/main/java/com/simpleAnalytics/TenetService/entity/Plan.java
@@ -20,7 +20,7 @@ public class Plan {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     UUID id;
     String name;
-    int monthlyCreditLimit;
+    long monthlyCreditLimit;
     int rateLimit;
     Date startDate;
     Duration duration;
